Remove merge_mining tag
Refactor conduit app to include error handling
Refactor conduit to include logging from tari base_node

rename conduit to merge mining proxy

remove partial_eq flag
