+++ Cargo.toml
@@ -16,5 +16,5 @@ members = [
     "applications/tari_base_node",
     "applications/tari_console_wallet",
     "applications/test_faucet",
-    "applications/tari_conduit",
+    "applications/tari_merge_mining_proxy",
 ]
