+++ applications/tari_merge_mining_proxy/README.md
@@ -32,15 +32,14 @@ Start XMRig from the terminal using `./xmrig`
 
 ### Tari Conduit
 
-In `main.rs` of the Tari Conduit the following constants can be changed to reflect the above configuration:
+In `main.rs` of the Tari merge mining proxy the following constants can be changed to reflect the above configuration:
 ```Rust
 const MONEROD_URL: &str = "monero-stagenet.exan.tech:38081"; // To connect to the Monero Public Stagenet.
 const TARI_GRPC_URL: &str = "http://127.0.0.1:18142"; // To connect to the Tari Base Node.
 const LOCALHOST: &str = "127.0.0.1:7878"; // For XMRig to connect.
 ```
 
-Then run the Tari Conduit with `cargo run --bin tari_conduit`.
+Then run the Tari Conduit with `cargo run --bin tari_merge_mining_proxy`.
 
 Blocks that have been mined for Monero Stagenet can be viewed here:
 `https://monero-stagenet.exan.tech/`
-
