+++ applications/tari_merge_mining_proxy/build.rs
